Cleanup - move helper classes
