+++ Assessments.Transformation/PublishDynamicProperties.cs
@@ -6,20 +6,10 @@ using System.Threading.Tasks;
 using Assessments.Shared.Helpers;
 using System.Text.Json;
 using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
-using Assessments.Shared.Options;
 using Assessments.Transformation.Helpers;
-using AutoMapper;
 using Assessments.Mapping.RedlistSpecies;
-using Raven.Client.Document;
+using Assessments.Transformation.DynamicProperties;
 using Raven.Client;
-using System;
-using Raven.Abstractions.Commands;
-using static Assessments.Mapping.RedlistSpecies.Source.Rodliste2019;
-using Raven.Abstractions.Extensions;
-using System.Diagnostics.CodeAnalysis;
-using System.Collections;
 
 namespace Assessments.Transformation
 {
@@ -32,8 +22,7 @@ namespace Assessments.Transformation
             _configuration = configuration;
             //var loggerFactory = LoggerFactory.Create(builder => builder.AddConsole());
         }
-
-
+        
         public async Task<List<DynamicProperty>> ImportRedlist2021()
         {
             var fileContent = await Storage.DownloadFromBlob(_configuration, DataFilenames.Species2021);
@@ -63,8 +52,7 @@ namespace Assessments.Transformation
             }).
             ToList();
         }
-
-
+        
         public async Task<List<DynamicProperty>> ImportAlienList2023()
         {
             var fileContent = await Storage.DownloadFromBlob(_configuration, DataFilenames.AlienSpecies2023);
@@ -126,72 +114,6 @@ namespace Assessments.Transformation
             StoreDynamicProperties(NewOrChangedDynamicProperties, ravenSession);
 
             ravenSession.SaveChanges();
-        }
-
-        public class DynamicPropertyIDComparer : IEqualityComparer<DynamicProperty>
-        {
-            public bool Equals(DynamicProperty x, DynamicProperty y)
-            {
-                return x.Id == y.Id;
-            }
-
-            public int GetHashCode([DisallowNull] DynamicProperty obj)
-            {
-                if (obj == null) { return 0; }
-
-                return obj.Id.GetHashCode();
-            }
-        }
-
-        public class DynamicPropertyObjectComparer : IEqualityComparer<DynamicProperty>
-        {
-            public bool Equals(DynamicProperty x, DynamicProperty y)
-            {
-                // If both are null, or both are the same instance, return true
-                if (ReferenceEquals(x, y))
-                {
-                    return true;
-                }
-
-                // If one is null, return false
-                if (x is null || y is null)
-                {
-                    return false;
-                }
-
-                return x.Id == y.Id &&
-                    ((IStructuralEquatable)x.References).Equals(y.References, EqualityComparer<string>.Default) &&
-                    CheckPropertiesEquality(x.Properties, y.Properties);
-            }
-
-            public int GetHashCode([DisallowNull] DynamicProperty obj)
-            {
-                if (obj == null) { return 0; }
-
-                int hashId = obj.Id.GetHashCode();
-                int hashReferences = ((IStructuralEquatable)obj.References).GetHashCode(EqualityComparer<string>.Default);                
-                int HashProperties = obj.Properties?.Select(p => p.Name?.GetHashCode() + p.Value?.GetHashCode() + p.Properties?.Select(p2 => p2.Name?.GetHashCode() + p2.Value?.GetHashCode() ?? 0).Sum()).Sum() ?? 0; //NOTE: calculates hash codes 2 levels down into the recurrent Property object, this should be sufficient for assessments
-
-                return HashCode.Combine(hashId, hashReferences, HashProperties);                
-            }
-        }
-
-        private static bool CheckPropertiesEquality(DynamicProperty.Property[] xProps, DynamicProperty.Property[] yProps)
-        {
-            if (xProps == yProps) return true;
-            if (xProps is null || yProps is null) return false;
-            if (xProps.Length != yProps.Length) return false;
-
-            for (int i = 0; i < xProps.Length; i++)
-            {
-                if (xProps[i].Name != yProps[i].Name || xProps[i].Value != yProps[i].Value)
-                    return false;
-
-                if (!CheckPropertiesEquality(xProps[i].Properties, yProps[i].Properties))
-                    return false;
-            }
-
-            return true;
         }
 
         private static void DeleteDynamicProperties(IDocumentSession ravenSession, List<DynamicProperty> obsoletDynamicProperties)
